refactor: adjusting architecture pattern
