+++ src/reducers/categories.js
